+++ src/main/java/de/maxhenkel/corpse/Config.java
@@ -24,25 +24,25 @@ public class Config {
     public static class ServerConfig {
         public ForgeConfigSpec.BooleanValue onlyOwnerAccess;
         public ForgeConfigSpec.IntValue corpseDespawnTime;
+        public ForgeConfigSpec.IntValue corpseForceDespawnTime;
         public ForgeConfigSpec.IntValue corpseSkeletonTime;
         public ForgeConfigSpec.BooleanValue spawnCorpseOnFace;
 
         public ServerConfig(ForgeConfigSpec.Builder builder) {
             onlyOwnerAccess = builder
                     .comment("If only the owner of the corpse can access the inventory")
-                    .translation("only_owner_access")
                     .define("only_owner_access", false);
             corpseDespawnTime = builder
                     .comment("The time passed after a corpse despawns (only if empty)")
-                    .translation("corpse_despawn_time")
                     .defineInRange("corpse_despawn_time", 20 * 30, 20, Integer.MAX_VALUE);
+            corpseForceDespawnTime = builder
+                    .comment("The time passed after a corpse despawns even if its not empty (-1 = never)")
+                    .defineInRange("corpse_force_despawn_time", -1, -1, Integer.MAX_VALUE);
             corpseSkeletonTime = builder
                     .comment("The time passed after a corpse turns into a skeleton")
-                    .translation("corpse_skeleton_time")
                     .defineInRange("corpse_skeleton_time", 20 * 60 * 60, 0, Integer.MAX_VALUE);
             spawnCorpseOnFace = builder
                     .comment("If the corpse should spawn on its face")
-                    .translation("spawn_corpse_on_face")
                     .define("spawn_corpse_on_face", false);
         }
     }